# public/swagger.json
{
  "info": {
    "title": "SHOP : API Documentation",
    "version": "1.0",
    "contact": "admin@spericorn.com",
    "description": "There are the api documentation for the student management system"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "swagger": "2.0",
  "paths": {
    "/api/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Signup for user",
        "operationId": "signup",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "signup for users",
            "required": true,
            "schema": {
              "$ref": "#/definitions/user"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Login for user after register",
        "parameters": [
          {
            "in": "body",
            "name": "user",
            "schema": {
              "type": "body",
              "properties": {
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/password/forgot": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Forgot password",
        "parameters": [
          {
            "in": "body",
            "name": "user",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/password/reset/{token}": {
      "put": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "Authentication"
        ],
        "summary": "Reset Password",
        "description": "Reset Password",
        "operationId": "ResetPassword",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "description": "token provided by email link thorugh (click the link and copy id form url)",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "type": "string"
                }
              }
            }
          },
          {
            "in": "body",
            "name": "body",
            "description": "enter your new password",
            "required": "on",
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "type": "string"
                },
                "confirmPassword": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "requrestBody": {
          "description": "request body",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "password": {
                    "type": "string"
                  },
                  "confirmPassword": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/logout": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "security": {
          "-Bearer": []
        },
        "summary": "Logout User",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/me": {
      "get": {
        "tags": [
          "User"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "view you profile",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/me/update": {
      "put": {
        "tags": [
          "User"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Update Profile",
        "description": null,
        "operationId": "update",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Update Your Profile",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/password/update": {
      "put": {
        "tags": [
          "User"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Update Password After Login",
        "description": null,
        "operationId": "update",
        "consumes": [
          "application/json",
          "applicaiton/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Update Your Password",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "oldPassword": {
                  "type": "string"
                },
                "newPassword": {
                  "type": "string"
                },
                "confirmPassword": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/user/{id}": {
      "put": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "Admin"
        ],
        "summary": "Update Role",
        "description": "Update Role",
        "operationId": "Update Role",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Token provided by email link through (Click the link and copy id from url)",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                }
              }
            }
          },
          {
            "in": "body",
            "name": "body",
            "description": "Enter Your New Password",
            "required": "on",
            "schema": {
              "type": "object",
              "properties": {
                "role": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "requestBody": {
          "description": "request body",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": null,
                "role": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "view single user",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "id",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Delete single user",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "id",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/users": {
      "get": {
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "View all users",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "definitions": {
    "user": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    }
  },
...
}